adding service object
